Struct static_ref::Static

#[repr(C)]

pub struct Static<T> where
    T: ?Sized,  { /* fields omitted */ }

A value stored in a static variable

Methods

impl<T> Static<T>

pub const unsafe fn new(value: T) -> Self

Asserts that this value is stored in a static variable

pub unsafe fn ref_<'a>(r: &'a T) -> &'a Static<T>

Asserts that the reference r points to data stored in a static variable

pub unsafe fn ref_mut<'a>(r: &'a mut T) -> &'a mut Static<T>

Asserts that the reference r points to data stored in a static variable

pub fn wrap(r: &'static T) -> &'static Static<T>

Wraps a static reference into Static

pub fn wrap_mut(r: &'static mut T) -> &'static mut Static<T>

Wraps a mutable static reference into Static